float64 as well in stress tracker

+++ alphafold2_pytorch/utils.py
@@ -332,7 +332,7 @@ def mds_torch(pre_dist_mat, weights=None, iters=10, tol=1e-5, verbose=2):
     diag_idxs = np.arange(N)
     his = []
     # init random coords
-    best_stress = float("Inf") * torch.ones(batch, device = device)
+    best_stress = float("Inf") * torch.ones(batch, device = device).type(dtype)
     best_3d_coords = 2*torch.rand(batch, N, 3, device = device).type(dtype) - 1
     # iterative updates:
     for i in range(iters):
